+++ flink-connector-elasticsearch8/src/test/java/org/apache/flink/connector/elasticsearch/sink/ElasticsearchSinkBaseITCase.java
@@ -22,6 +22,9 @@
 package org.apache.flink.connector.elasticsearch.sink;
 
 import org.apache.http.HttpHost;
+import org.apache.http.util.EntityUtils;
+import org.elasticsearch.client.Request;
+import org.elasticsearch.client.Response;
 import org.elasticsearch.client.RestClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,10 +33,12 @@ import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.elasticsearch.ElasticsearchContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
 
+import java.io.IOException;
 import java.time.Duration;
 
-import static org.apache.flink.connector.elasticsearch.sink.Elasticsearch8TestUtils.ELASTICSEARCH_IMAGE;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /** Base Integration tests class. */
 @Testcontainers
@@ -41,6 +46,12 @@ public class ElasticsearchSinkBaseITCase {
     protected static final Logger LOG = LoggerFactory.getLogger(ElasticsearchSinkBaseITCase.class);
 
     public RestClient client;
+
+    public static final String ELASTICSEARCH_VERSION = "8.12.1";
+
+    public static final DockerImageName ELASTICSEARCH_IMAGE =
+            DockerImageName.parse("docker.elastic.co/elasticsearch/elasticsearch")
+                    .withTag(ELASTICSEARCH_VERSION);
 
     @Container
     public static final ElasticsearchContainer ES_CONTAINER = createElasticsearchContainer();
@@ -66,5 +77,52 @@ public class ElasticsearchSinkBaseITCase {
                         new HttpHost(ES_CONTAINER.getHost(), ES_CONTAINER.getFirstMappedPort()))
                 .setHttpClientConfigCallback(httpClientBuilder -> httpClientBuilder)
                 .build();
+    }
+
+    public static void assertIdsAreWritten(RestClient client, String index, String[] ids)
+            throws IOException {
+        client.performRequest(new Request("GET", "_refresh"));
+        Response response = client.performRequest(new Request("GET", index + "/_search/"));
+        String responseEntity = EntityUtils.toString(response.getEntity());
+
+        assertThat(response.getStatusLine().getStatusCode()).isEqualTo(200);
+
+        for (String id : ids) {
+            System.out.println(id);
+            assertThat(responseEntity).contains(id);
+        }
+    }
+
+    public static void assertIdsAreNotWritten(RestClient client, String index, String[] ids)
+            throws IOException {
+        client.performRequest(new Request("GET", "_refresh"));
+        Response response = client.performRequest(new Request("GET", index + "/_search/"));
+        String responseEntity = EntityUtils.toString(response.getEntity());
+
+        assertThat(response.getStatusLine().getStatusCode()).isEqualTo(200);
+
+        for (String id : ids) {
+            assertThat(responseEntity).doesNotContain(id);
+        }
+    }
+
+    /** DummyData is a POJO to helping during integration tests. */
+    public static class DummyData {
+        private final String id;
+
+        private final String name;
+
+        public DummyData(String id, String name) {
+            this.id = id;
+            this.name = name;
+        }
+
+        public String getId() {
+            return id;
+        }
+
+        public String getName() {
+            return name;
+        }
     }
 }
